+++ components-ui/app/src/tol-ui/src/index.ts
@@ -46,6 +46,7 @@ export * from "./hooks";
 export * from "./map";
 export * from "./messaging";
 export * from "./interfaces";
+export * from "./links";
 export * from "./overlays";
 export * from "./relationships";
 export * from "./services";